wip: Modify function for deleting flows.
When the dump_duration is greater than 500ms and we are using >50% of
the flow_limit for flows, start deleting flows.

                                        n_flows / flow_limit
idle_age | % of max_idle |     60% |    75% |    80% |   90%  |   100%
---------+---------------+---------+--------+--------+--------+--------
   100ms |           ~7% |    0.5% |   1.1% |   1.3% |   1.6% |   2.2%
   750ms |           50% |    3.3% |   8.3% |  10.0% |  12.5% |  16.6%
  1500ms |          100% |    6.6% |  16.6% |  20.0% |  25.0% |  33.3%

Note that we are using modulo for deleting flows, so it is more stepped
than this table may appear. At max_idle time, at the flow limit this
deletes as (flow_hash % 3).

This appears to be more stable on the limit, and increases TCP CRR
performance by up to 4%. However, in netmap test this decreases the
number of flows we can support in the datapath. For instance, running
5:2 threads with 1500ms idle time, spamming 64K ports, drops n_flows
from ~70K->~40K.

Signed-off-by: Joe Stringer <joestringer@nicira.com>