wip: Introduce linear function for deleting flows.
When the dump_duration is greater than 500ms and we are using >70% of
the flow_limit for flows, start deleting flows (but never >50%).

                                   n_flows / flow_limit
idle_age | % of max_idle |     75% |    85% |    90% |   100%
---------+---------------+---------+--------+--------+--------
   100ms |           ~7% |    0.5% |   1.3% |   2.2% |   3.3%
   750ms |           50% |    3.5% |  10.0% |  16.6% |  25.0%
  1500ms |          100% |    7.1% |  20.0% |  33.3% |  50.0%

Note that we are using modulo for deleting flows, so it is more stepped
than this table may appear. At max_idle time, at the flow limit this
deletes as (flow_hash % 2).

Signed-off-by: Joe Stringer <joestringer@nicira.com>