revalidator: Selectively perform xlate side-effects.
Commit a48c85b2d6 (revalidator: Use xcache when revalidation is
required.) introduced a bug where xlate_actions effects from the old
cache would be performed. This could cause stale mac-learning entries to
re-appear, and in some cases, learnt flows to be re-learnt even after
the parent flow is deleted.

This patch splits the attribution of stats from the execution of side
effects for the flow, always attributing the stats to the rules that
created them, but selectively executing side-effects based on the
validity of the flow.

This handles three cases in slightly different ways:
* If revalidation is unnecessary, attribute stats and perform
  side-effects using the existing xcache.
* If revalidation is necessary, and the flow is valid, then attribute
  stats using the existing xcache, then perform full revalidation.
  Perform side-effects based on the newly built xcache.
* If revalidation is necessary, and the flow is invalid, then attribute
  stats using the old xcache, and do not perform side-effects.

VMware-BZ: #1268574

Reported-by: Len Gao <leng@vmware.com>
Signed-off-by: Joe Stringer <joestringer@nicira.com>