ovs-dev.py: Add command to generate coverage info.
This will generate HTML output of the testsuite coverage in
_build-gcc/coverage/*.

After running "ovs-dev.py coverage", the source will be configured to
collect coverage information. This may cause testsuite failures,
particularly if running with high parallelism. Re-run "ovs-dev.py conf"
to remove the coverage configuration.

Signed-off-by: Joe Stringer <joestringer@nicira.com>
---
I've uploaded an initial run of the test coverage here for a sample:
http://joestringer.github.io/openvswitch/index.html

If people are interested in having a coverage site like this accessible
all the time, I could probably set up a cron job to keep it up to date.

The main thing I've noticed so far is that IPFIX is almost completely
untested. There are some limitations, this is not running the linux
datapath and attached netlink code. Furthermore the testsuite primarily
tests for true positives/false negatives, and doesn't check for false
positives/true negatives. This decreases the % coverage considerably.