upcall: Create ukeys in handler threads.
Currently, when a revalidator thread first dumps a flow, it creates a
'udpif_key' object and caches a copy of a kernel flow key. This allows
us to perform lookups in the classifier to attribute stats and validate
the correctness of the datapath flow.

This patch sets up this cache from the handler threads, during flow
setup. This allows future patches to reduce the cost of flow dumping.

Signed-off-by: Joe Stringer <joestringer@nicira.com>
---
v3: Rebase.
v2: Rebase.
RFC: First post.