dpif: Add Unique flow identifiers.
One of the limiting factors on the number of flows that can be supported
in the datapath is the overhead of assembling flow dump messages in the
datapath. This patch adds a new alternative to dumping the key, mask and
actions from the datapath, which is a 128-bit unique identifier for the
flow. For each flow dump, the dpif user specifies which fields may be
omitted, allowing the common case to only dump a pair of 128-bit ID and
flow stats.

Signed-off-by: Joe Stringer <joestringer@nicira.com>
---
I'm still looking at whether it might be worth trying a 64-bit UID
instead. Any feedback on this would be welcome.

v3: Rebase.