upcall: Generate unique flow identifiers.
This patch modifies the way that ukey lookup is done, to use a 128-bit
hash of the flow key as Netlink attributes. This hash is generated at
flow installation time, and passed down to the dpif so that datapaths
which support UID can index based on this ID rather than the flow.

For backward compatibility, flows that are dumped with no ukey will be
hashed in the same manner as flow setup, so that the flow may be looked
up. For this reason, it is important that two flows with the same UID
are not installed. Handlers enforce this and track conflicts with the
coverage counter 'handler_install_conflict'.

Signed-off-by: Joe Stringer <joestringer@nicira.com>
---
v4: Fix bug where UID-based terse dump wasn't enabled by default.
    Display whether terse dump is enabled in ovs-appctl upcall/show.
v3: Rename the ovs-appctl commands from {en,dis}able-uid->*-terse-dump.
    Rebase.
v2: No change.