hack: Fix dp_netdev ukey install/flow install race.
There's a race where one thread installs a ukey, then a revalidator
thread completes an entire cycle of seq increment, flow dump and sweep
before the first thread installs the flow into the datapath. This will
cause the revalidator thread to fail to get the flow.

Hack together a fix for this race.

Signed-off-by: Joe Stringer <joestringer@nicira.com>