dpif: Extract 'ufid' into its own struct.
Previously we've had a separation between the unique flow identifier
(ufid) and an additional bool 'ufid_present' which tracks whether that
particular flow was installed using UFID (which is typically true if the
datapath supports the feature). Then, in ofproto-dpif-upcall when making
calls into the dpif API, it would decide whether to pass the UFID down
depending on this value.

If we factor these out into a 'struct ufid', then only one place in
ofproto-dpif-upcall needs to care about the status of the bool: When
deciding to install a flow, we determine backer support for UFID and set
the 'ufid.propagate' bit. All other places can blindly pass around the
'struct ufid' and the bit will be in the correct state, then the dpif
can decide whether it should forward the UFID to the datapath or not.

Signed-off-by: Joe Stringer <joe@ovn.org>