compat: Fix fragmentation issue with tunnels.
Prior to upstream Linux commit d6b915e29f4a ("ip_fragment: don't forward
defragmented DF packet"), the fragmentation behaviour was incorrect when
dealing with linear skbs, as it would not respect the "max_frag_size"
that ip_defrag() provides, but instead attempt to use the output
device's MTU.

If OVS reassembles an IP message and passes it up to userspace, it
linearizes the data to pass up, and also provides a PACKET_ATTR_MRU to
indicate the maximum received unit size for this message. When userspace
executes actions to output this packet, it passes the MRU back down and
this is the desired refragment size. However, due to the above bug it
would use the device's MTU to refragment. In the case of regular ports,
this is not too dangerous as the MTU would be a reasonable value.
However, in the case of a tunnel port the typical MTU is a very large
value. As such, rather than refragmenting the message on output, it
would simply output the (too-large) frame to the tunnel.

Depending on the tunnel type and other factors, this large frame could
be dropped along the path, or it could end up at the remote tunnel
endpoint and end up being delivered towards a remote host stack or VM.
If OVS is also controlling that endpoint, it will likely drop the packet
when sending to the final destination, because the packet exceeds the
port MTU.

Different OpenFlow rule configurations could end up preventing IP
messages from being refragmented correctly for as many as the first four
attempts in each connection.

Fix this issue by backporting ip_do_fragment() so that it will respect
the MRU value that is provided in the execute path.

VMWare-BZ: #1651589
Fixes: 213e1f54b4b3 ("compat: Wrap IPv4 fragmentation.")
Reported-by: Salman Malik <salmanm@vmware.com>
Signed-off-by: Joe Stringer <joe@ovn.org>
---
This patch is verified by running the kernel module test "conntrack -
Fragmentation over vxlan". Unfortunately in my previous setups I was
unable to notice this because the test is skipped when iproute2 is an
older version, which is the case on many common platforms. The bug only
affects the backport for kernels 4.1 and older.