ofproto-dpif-upcall: Replace ukeys for deleted flows.
If a revalidator dumps/revalidates a flow during the 'dump' phase,
resulting in the deletion of the flow, then ukey->flow_exists is set to
false, and the ukey is kept around until the 'sweep' phase. The ukey is
kept around to ensure that cases like duplicated dumps from the
datapaths do not result in multiple attribution of the same stats.

However, if an upcall for this flow comes for a handler between the
revalidator 'dump' and 'sweep' phases, the handler will lookup the ukey
and find that the ukey exists, then skip installing a new flow entirely.
As a result, for this period all traffic for the flow is slowpathed.
If there is a lot of traffic hitting this flow, then it will all be
handled in userspace until the 'sweep' phase. Eventually the
revalidators will reach the sweep phase and delete the ukey, and
subsequently the handlers should install a new flow.

To reduce the slowpathing of this traffic during flow table transitions,
allow the handler to identify this case during miss upcall handling and
replace the existing ukey with a new ukey. The handler will then be able
to install a flow for this traffic, allowing the traffic flow to return
to the fastpath.

There are three situations where the existing ukey has "flow_exists" set
to false:

* Firstly, the flow for the ukey might not yet have been installed. In
  this case, the other handler that set up this ukey will be holding the
  lock for the ukey. So, if we attempt to grab the lock and fail then
  another handler is already setting up the flow and we can safely skip
  flow install.
* Secondly, a revalidator could be currently deleting the flow. In this
  case, the revalidator holds the ukey lock so the handler will fail to
  grab it. This is fine, if traffic continues to miss then a subsequent
  miss upcall will hit the third case.
* Thirdly, the flow may have been recently deleted by a revalidator
  thread. In this case, we can grab the lock. From the handler thread
  we swap the original key out for a new one and rcu-defer its deletion.

Signed-off-by: Joe Stringer <joe@ovn.org>