dpif-hw-acc: Add new hash maps ufid <-> tc flow and ovs port <-> netdev
A flow in ovs can be identified by its ufid, and in tc by its
flow handle + priority.
We use the flow protocol number as the priority when offloading to tc.
So we save a map that maps from ovs ufid <-> tc flow handle + protocol.
Also added a map to map ovs in_port to/from netdevice to be used
when offloading/dumping a flow.

Signed-off-by: Paul Blakey <paulb@mellanox.com>
Signed-off-by: Shahar Klein <shahark@mellanox.com>