configure: Reset libtool CURRENT version.
Since commit XXXXXXXXXXXX ("libopenvswitch: Rename to libfoo-X.Y."), the
CURRENT libtool number is no longer derived from the OVS MINOR (from
vMAJOR.MINOR.MICRO) version, so it can be reset to 0.

Developers should attempt to avoid introducing ABI-breaking changes
within a particular OVS-X.Y release series. Occasionally due to the
nature of a particular bug, this is not possible. In such a case,
developers must update the libtool CURRENT version to indicate this
breakage to library users.

In most OVS library releases, this is expected to remain 0.

Signed-off-by: Joe Stringer <joe@ovn.org>