tunnel-tests: Add test to match tunnel traffic.
This test highlights a bug that was affecting master up until the
previous patch. Put simply, we have two bridges: an integration bridge
which contains a tunnel, and a physical bridge for underlay network
connectivity. This test simulates putting UDP traffic through the
integration bridge, with the intention to apply GRE tunnel headers and
send the packet through the underlay bridge. The underlay bridge should
observe GRE traffic.