odp-execute: Fix unaligned eth_addr pointers.
Clang 4.0 complains:

../lib/odp-execute.c:61:37: error: taking address of packed member 'eth_dst' of
class or structure 'eth_header' may result in an unaligned pointer value
      [-Werror,-Waddress-of-packed-member]
            ether_addr_copy_masked(&eh->eth_src, key->eth_src, mask->eth_src);
                                    ^~~~~~~~~~~
../lib/odp-execute.c:62:37: error: taking address of packed member 'eth_dst' of
class or structure 'eth_header' may result in an unaligned pointer value
      [-Werror,-Waddress-of-packed-member]
            ether_addr_copy_masked(&eh->eth_dst, key->eth_dst, mask->eth_dst);

Ethernet source addresses are 48 bits offset into the Ethernet header,
so taking a pointer for this is not guaranteed to be valid on all
architectures. Fix this by referencing the memory direct from the
Ethernet header pointer.

Signed-off-by: Joe Stringer <joe@ovn.org>