XXX: Splitme
Currently the stubs for if a DPIF doesn't support meters are all
implemented in dpif-netlink. However, it makes sense instead for this
stub logic to be put in the DPIF layer and just set the pointers to NULL
in dpif-netlink. This simplifies stubbing out meter support in a new
upcoming dpif.

XXX: CT action notes:
    Make use of the cilium-derived connection tracker introduced in an
    earlier commit to add support for connection tracking in the BPF
    datapath.

    This patch doesn't actually modify the metadata attached to the flow,
    because apparently we use this metadata to perform our next-action
    lookup so if we modify it then we can't find the next action in our
    list! This is analogous to having recirc() executed after every action.
    We need to fix this issue before really addressing connection tracking
    in a meaningful way.

    TODO:
    * Double-check that ingress/egress is treated correctly
    * Zones
    * Mark
    * Labels
    * Force commit
    * NAT
    * Helpers
    * Eventmasks

Signed-off-by: Joe Stringer <joe@ovn.org>