dpif: Refactor stub implementation for meter_*().
Currently the stubs for if a DPIF doesn't support meters are all
implemented in dpif-netlink. However, it makes sense instead for this
stub logic to be put in the DPIF layer and just set the pointers to NULL
in dpif-netlink. This simplifies stubbing out meter support in a new
upcoming dpif.

Signed-off-by: Joe Stringer <joe@ovn.org>