XXX: bpf: Add CT action
Make use of the cilium-derived connection tracker introduced in an
earlier commit to add support for connection tracking in the BPF
datapath.

This patch doesn't actually modify the metadata attached to the flow,
because apparently we use this metadata to perform our next-action
lookup so if we modify it then we can't find the next action in our
list! This is analogous to having recirc() executed after every action.
We need to fix this issue before really addressing connection tracking
in a meaningful way.

TODO:
* Double-check that ingress/egress is treated correctly
* Zones
* Mark
* Labels
* Force commit
* NAT
* Helpers
* Eventmasks