bpf: Mock out cilium APIs necessary for compile.
The previous commit imported a bunch of cilium code for connection
tracking, but it won't compile in the tree at the moment. Fortunately
we're not trying to compile it just yet. This commit will make a few
modifications to the aforementioned library to allow it to compile and
be used within the OVS BPF datapath.